fixed on VS2022 to be compiled

+++ FRHED/hgstream.cpp
@@ -55,7 +55,7 @@ HGlobalStream& HGlobalStream::operator << (BYTE c)
 {
 	if (_hex)
 	{
-		BYTE nib[2] = { c >> 4 & 0xf, c & 0xf };
+		BYTE nib[2] = { c >> 4 & (BYTE)0xf, c & (BYTE)0xf };
 		nib[0] += nib[0] >= 0xa ? 'a' - 0xa : '0';
 		nib[1] += nib[1] >= 0xa ? 'a' - 0xa : '0';
 		Realloc(sizeof nib, nib);
